Only remove the params is preserve=0 and not also on 1

+++ src/Controller/Component/QueryParamPreserverComponent.php
@@ -120,7 +120,7 @@ class QueryParamPreserverComponent extends Component
         $params = $request->getQueryParams();
         $ignoreParam = $this->getConfig('disablePreserveWithParam');
 
-        if (isset($params[$ignoreParam])) {
+        if (isset($params[$ignoreParam]) && (bool)$params[$ignoreParam] === false) {
             unset($params[$ignoreParam]);
 
             $request->session()->delete($this->_hashKey());
